+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.2.8]
+
+## Fix
+- fixed bug for single entries in intvlsize
+
+# [0.2.7]
+
+## Fix
+- fixed bug with missing BenchTabix class
+
 # [0.2.6]
 
 ## Fix